Move internal sub_module
