+++ src/snowflake/snowpark/_functions/__init__.py
@@ -1,0 +1,3 @@
+#
+# Copyright (c) 2012-2025 Snowflake Computing Inc. All rights reserved.
+#
