+++ src/snowflake/snowpark/functions.py
@@ -168,16 +168,16 @@ from typing import Callable, Dict, List, Optional, Tuple, Union, overload
 
 import snowflake.snowpark
 import snowflake.snowpark._internal.proto.generated.ast_pb2 as proto
-from snowflake.snowpark._functions.general_functions import (
+from snowflake.snowpark._internal.general_functions import (
     _call_function,
     _check_column_parameters,
     lit,
     call_function,
     function,
     builtin,
+    col as col_function,
 )
 
-from snowflake.snowpark._functions.general_functions import col as col_function
 
 import snowflake.snowpark.table_function
 from snowflake.snowpark._internal.analyzer.expression import (
